remove lcfirst, use first media file
removed the lcfirst lowercase first letter. make this function optional so user can decide or delete it...
users how nee the first letter in lowercase should fix/update their gedcom data

change the preg match for the OBJE so only Level 1 Mediafiles are used, and the first image is most likely the wanted profile picture
without this change i also got documents displayed which are not wanted in the report.

feel free to close this PR if you dont agree :)

Signed-off-by: Holger Eisold <github@eisold.family>